cpu/pic32: Configure prefetch cache
The number of wait states to access program memory is by
default 7. This number can be reduced to only 2 if the system
clock is at 80MHz.
This commit computes the minimum number of wait states required
and enables predictive prefetching for all regions. It also
reserves one cache line for data.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>