cpu/pic32: Ensure that TX fifo is flushed before disabling UART
Disabling UART peripheral also flushes its FIFO. Since we do not
wait until each character is transmitted in the write function,
some characters could be dropped.

This commit uses the TXIF flag to check if the TX FIFO is empty.
It waits that the FIFO becomes empty when releasing and powering
down UART.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>